// src/clients/generated/commonApi/models/uiOrderItemPreparationState.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */

/**
 * Статус готовности изделия.
 */
export type UiOrderItemPreparationState = typeof UiOrderItemPreparationState[keyof typeof UiOrderItemPreparationState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UiOrderItemPreparationState = {
  NOT_STARTED: 'NOT_STARTED',
  IN_PROCESS: 'IN_PROCESS',
  DONE: 'DONE',
} as const;